backend/requirements.txt
Django==5.2.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
python-decouple==3.8
argon2-cffi==23.1.0
pyotp==2.9.0
qrcode[pil]==7.4.2
Pillow==11.0.0

# Optional: PostgreSQL support (uncomment if using PostgreSQL)
# psycopg2-binary==2.9.9

# Optional: Redis support (uncomment if using Redis)
# redis==5.0.1
# django-redis==5.4.0